Clean up and refactor extension classes Fix bugs in localisation table
