+++ src/Extension/Traits/FluentBadgeTrait.php
@@ -1,84 +1,49 @@
 <?php
 
-namespace TractorCow\Fluent\Extension;
+namespace TractorCow\Fluent\Extension\Traits;
 
-use SilverStripe\Core\Extension;
-use SilverStripe\ORM\ArrayList;
 use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\FieldType\DBField;
+use TractorCow\Fluent\Extension\FluentExtension;
 use TractorCow\Fluent\Model\Locale;
-use TractorCow\Fluent\State\FluentState;
 
-class FluentBadgeExtension extends Extension
+trait FluentBadgeTrait
 {
-    /**
-     * Push a badge to indicate the language that owns the current item
-     *
-     * @param DBField|null $badgeField
-     */
-    public function updateBadge(&$badgeField)
-    {
-        /** @var DataObject $record */
-        $record = $this->owner->getRecord();
-
-        $badgeField = $this->addFluentBadge($badgeField, $record);
-    }
-
-    public function updateBreadcrumbs(ArrayList $breadcrumbs)
-    {
-        // Ensure this doesn't run when it can't get the correct record
-        if (!$this->owner->hasMethod('currentPage')) {
-            return;
-        }
-
-        /** @var DataObject $record */
-        $record = $this->owner->currentPage();
-        if (!$record) {
-            return;
-        }
-
-        // Get a possibly existing badge field from the last item in the breadcrumbs list
-        $lastItem = $breadcrumbs->last();
-        $badgeField = $lastItem->hasField('Extra') ? $lastItem->getField('Extra') : null;
-        $newBadge = $this->addFluentBadge($badgeField, $record);
-
-        $lastItem->setField('Extra', $newBadge);
-    }
 
     /**
      * Add the Fluent state badge before any existing badges and return the result
      *
-     * @param DBField|null $badgeField
-     * @param DataObject $record
+     * @param DBField|null $badgeField Existing badge to merge with
+     * @param DataObject   $record
      * @return DBField|null
      */
     protected function addFluentBadge($badgeField, DataObject $record)
     {
-        // Check for a fluent state before continuing
-        if (!FluentState::singleton()->getLocale()
-            || !$record->has_extension(FluentVersionedExtension::class)
-        ) {
+        $fluentBadge = $this->getBadge($record);
+        if (!$fluentBadge) {
             return $badgeField;
         }
 
-        $fluentBadge = $this->getBadge($record);
         // Add fluent badge before any existing badges
         $newBadge = DBField::create_field('HTMLFragment', $fluentBadge . $badgeField);
-
         return $newBadge;
     }
 
     /**
      * Given a record with Fluent enabled, return a badge that represents the state of it in the current locale
      *
-     * @param DataObject&FluentVersionedExtension $record
-     * @return DBField
+     * @param DataObject|FluentExtension $record
+     * @return DBField|null
      */
     public function getBadge(DataObject $record)
     {
         /** @var Locale $currentLocale */
         $currentLocale = Locale::getCurrentLocale();
+        if (!$currentLocale || !$record->has_extension(FluentExtension::class)) {
+            return null;
+        }
 
+        // Bulid new badge
         $badgeClasses = ['badge', 'fluent-badge'];
         if ($currentLocale->getIsDefault()) {
             // Current locale should always show a "default" or green state badge
